Cleanup + Fix Storybook
