+++ .storybook/components/Explore/Explore.stories.tsx
@@ -1,4 +1,4 @@
-import { ExploreEventsView } from "@screens/ExploreEvents"
+import { ExploreEventsView } from "explore-events"
 import { EventMocks } from "@event-details/MockData"
 import { ComponentMeta, ComponentStory } from "@storybook/react-native"
 import { TiFQueryClientProvider } from "@lib/ReactQuery"
@@ -12,10 +12,6 @@ import { createStackNavigator } from "@react-navigation/stack"
 import { BASE_HEADER_SCREEN_OPTIONS } from "@components/Navigation"
 import { NavigationContainer } from "@react-navigation/native"
 import { SafeAreaProvider } from "react-native-safe-area-context"
-import {
-  getCurrentPositionAsync,
-  requestForegroundPermissionsAsync
-} from "expo-location"
 
 const ExploreEventsMeta: ComponentMeta<typeof ExploreEventsView> = {
   title: "Explore Events Screen",
@@ -27,14 +23,11 @@ export default ExploreEventsMeta
 type ExploreEventsStory = ComponentStory<typeof ExploreEventsView>
 
 const Stack = createStackNavigator<ExploreEventsScreensParamsList>()
-const screens = createExploreEventsScreens(Stack, () => ({
-  value: Promise.resolve([
-    EventMocks.Multiday,
-    EventMocks.NoPlacemarkInfo,
-    EventMocks.PickupBasketball
-  ]),
-  cancel: () => {}
-}))
+const screens = createExploreEventsScreens(Stack, async () => [
+  EventMocks.Multiday,
+  EventMocks.NoPlacemarkInfo,
+  EventMocks.PickupBasketball
+])
 
 export const Basic: ExploreEventsStory = () => (
   <MenuProvider>
