+++ explore-events/ExploreEvents.tsx
@@ -1,9 +1,20 @@
-import React from "react"
+import React, { useState } from "react"
 import { Pressable, StyleProp, StyleSheet, View, ViewStyle } from "react-native"
 import { useSafeAreaInsets } from "react-native-safe-area-context"
-import { LocationCoordinate2D, Region } from "@location/index"
+import {
+  LocationCoordinate2D,
+  Region,
+  useRequestForegroundLocationPermissions,
+  useUserCoordinatesQuery
+} from "@location/index"
 import { CurrentUserEvent } from "@shared-models/Event"
-import { ExploreEventsData } from "./models"
+import {
+  ExploreEventsData,
+  ExploreEventsInitialCenter,
+  SAN_FRANCISCO_DEFAULT_REGION,
+  createDefaultMapRegion,
+  initialCenterToRegion
+} from "./Models"
 import { ExploreEventsMap } from "./Map"
 import { useLastDefinedValue } from "@lib/utils/UseLastDefinedValue"
 import { ExploreEventsBottomSheet } from "./BottomSheet"
@@ -12,6 +23,125 @@ import { SkeletonEventCard } from "./SkeletonEventCard"
 import { Ionicon } from "@components/common/Icons"
 import { PrimaryButton } from "@components/Buttons"
 import { ExploreEventsSearchBar } from "./SearchBar"
+import { QueryHookOptions } from "@lib/ReactQuery"
+import { eventDetailsQueryKey } from "@shared-models/query-keys/Event"
+import { UseQueryResult, useQueryClient, useQuery } from "@tanstack/react-query"
+import { PermissionResponse, LocationAccuracy } from "expo-location"
+
+export type UseExploreEventsEnvironment = {
+  fetchEvents: (
+    region: Region,
+    signal?: AbortSignal
+  ) => Promise<CurrentUserEvent[]>
+  isSignificantlyDifferentRegions: (r1: Region, r2: Region) => boolean
+}
+
+/**
+ * A hook to provide the data and current region for the event exploration view.
+ */
+export const useExploreEvents = (
+  initialCenter: ExploreEventsInitialCenter,
+  { fetchEvents, isSignificantlyDifferentRegions }: UseExploreEventsEnvironment
+) => {
+  const { region, panToRegion } = useExploreEventsRegion(initialCenter)
+  const { events, cancel } = useExploreEventsQuery(region!, fetchEvents, {
+    enabled: !!region
+  })
+  return {
+    region,
+    data: eventsQueryToExploreEventsData(events),
+    updateRegion: (newRegion: Region) => {
+      if (!region) {
+        panToRegion(newRegion)
+      } else if (isSignificantlyDifferentRegions(region, newRegion)) {
+        cancel()
+        setTimeout(() => panToRegion(newRegion), 300)
+      }
+    }
+  }
+}
+
+const eventsQueryToExploreEventsData = (
+  query: UseQueryResult<CurrentUserEvent[], unknown>
+): ExploreEventsData => {
+  if (query.isLoading) {
+    return { status: "loading" }
+  } else if (query.isSuccess && query.data.length === 0) {
+    return { status: "no-results", events: [] }
+  } else if (query.isError) {
+    return { status: "error", retry: query.refetch }
+  }
+  return { status: "success", events: query.data }
+}
+
+const useExploreEventsRegion = (initialCenter: ExploreEventsInitialCenter) => {
+  const [pannedRegion, setPannedRegion] = useState(
+    initialCenterToRegion(initialCenter)
+  )
+  const userRegion = useUserRegion({ enabled: !pannedRegion })
+  const exploreRegion =
+    userRegion === "loading"
+      ? pannedRegion
+      : pannedRegion ?? userRegion ?? SAN_FRANCISCO_DEFAULT_REGION
+  return { region: exploreRegion, panToRegion: setPannedRegion }
+}
+
+type UserRegionResult = "loading" | (Region | undefined)
+
+const useUserRegion = (
+  options: QueryHookOptions<PermissionResponse>
+): UserRegionResult => {
+  const permissionQuery = useRequestForegroundLocationPermissions(options)
+  const locationQuery = useUserCoordinatesQuery(
+    { accuracy: LocationAccuracy.Balanced },
+    {
+      enabled: permissionQuery.data !== undefined
+    }
+  )
+  if (permissionQuery.isFetching || locationQuery.isFetching) {
+    return "loading"
+  }
+  const coords = locationQuery.data ? locationQuery.data?.coords : undefined
+  return !coords
+    ? undefined
+    : createDefaultMapRegion({
+        latitude: coords.latitude,
+        longitude: coords.longitude
+      })
+}
+
+const useExploreEventsQuery = (
+  region: Region,
+  fetchEvents: (
+    region: Region,
+    signal?: AbortSignal
+  ) => Promise<CurrentUserEvent[]>,
+  options: QueryHookOptions<CurrentUserEvent[]>
+) => {
+  const queryClient = useQueryClient()
+  const queryKey = ["explore-events", region]
+  return {
+    events: useQuery(
+      queryKey,
+      async ({ signal }) => {
+        const events = await fetchEvents(region, signal)
+
+        events.forEach((event) => {
+          queryClient.setQueryData(eventDetailsQueryKey(event.id), {
+            status: "success",
+            event
+          })
+        })
+
+        return events
+      },
+      options
+    ),
+    cancel: () => {
+      queryClient.cancelQueries({ queryKey })
+    }
+  }
+}
 
 export type ExploreEventsProps = {
   searchText?: string
